refactor: type에 맞게 props 이름 변경
